test: split huge CreateOccurrencePage test file to multiple files
